+++ utils/server/allocatedWorkers.js
@@ -47,18 +47,10 @@ export const addAllocatedWorker = async (mosaicId, params) => {
   return data;
 };
 
-export const getAllocations = async (worker_id, params) => {
+export const getAllocationsByWorker = async (worker_id, params) => {
   const { data } = await axios.get(`${ENDPOINT_API}/allocations`, {
     headers: { 'x-api-key': AWS_KEY },
     params: { worker_id, ...params },
-  });
-  return data;
-};
-
-export const getWorker = async (id, params) => {
-  const { data } = await axios.get(`${ENDPOINT_API}/workers`, {
-    headers: { 'x-api-key': AWS_KEY },
-    params: { id, ...params },
   });
   return data;
 };
