+++ utils/server/allocatedWorkers.spec.js
@@ -146,12 +146,12 @@ describe('allocatedWorkersAPI', () => {
     });
   });
 
-  describe('getAllocations', () => {
+  describe('getAllocationsByWorker', () => {
     it('should work properly', async () => {
       axios.get.mockResolvedValue({
         data: { allocations: ['foo'] },
       });
-      const data = await allocatedWorkersAPI.getAllocations(123);
+      const data = await allocatedWorkersAPI.getAllocationsByWorker(123);
       expect(axios.get).toHaveBeenCalled();
       expect(axios.get.mock.calls[0][0]).toEqual(`${ENDPOINT_API}/allocations`);
       expect(axios.get.mock.calls[0][1].headers).toEqual({
@@ -159,24 +159,6 @@ describe('allocatedWorkersAPI', () => {
       });
       expect(axios.get.mock.calls[0][1].params).toEqual({
         worker_id: 123,
-      });
-      expect(data).toEqual({ allocations: ['foo'] });
-    });
-  });
-
-  describe('getWorker', () => {
-    it('should work properly', async () => {
-      axios.get.mockResolvedValue({
-        data: { allocations: ['foo'] },
-      });
-      const data = await allocatedWorkersAPI.getWorker(123);
-      expect(axios.get).toHaveBeenCalled();
-      expect(axios.get.mock.calls[0][0]).toEqual(`${ENDPOINT_API}/workers`);
-      expect(axios.get.mock.calls[0][1].headers).toEqual({
-        'x-api-key': AWS_KEY,
-      });
-      expect(axios.get.mock.calls[0][1].params).toEqual({
-        id: 123,
       });
       expect(data).toEqual({ allocations: ['foo'] });
     });
